// src/app/interfaces/measurement-form-data.ts
export interface MeasurementFormData{
  CTOTAL___: string;
  STOTAL__: string;
  SiO2: string;
  Al2O3: string;
  Fe2O3: string;
  MgO: string;
  CaO: string;
  Na2O: string;
  K2O: string;
  TiO2: string;
  P2O5: string;
  MnO: string;
  Cr2O3: string;
  LOI: string;
  Suma: string;
  Sc: string;
  Ba: string;
  Be: string;
  Co: string;
  Cs: string;
  Ga: string;
  Hf: string;
  Nb: string;
  Rb: string;
  Sn: string;
  Sr: string;
  Ta: string;
  Th: string;
  U: string;
  V: string;
  W: string;
  Zr: string;
  Y: string;
  La: string;
  Ce: string;
  Pr: string;
  Nd: string;
  Sm: string;
  Eu: string;
  Gd: string;
  Tb: string;
  Dy: string;
  Ho: string;
  Er: string;
  Tm: string;
  Yb: string;
  Lu: string;
  Mo: string;
  Cu: string;
  Pb: string;
  Zn: string;
  Ni: string;
  As_: string;
  Cd: string;
  Sb: string;
  Bi: string;
  Ag: string;
  Au: string;
  Hg: string;
  Tl: string;
  Se: string;
  Te_ppm: string;
  B_ppm: string;
  geochemistry_published: boolean;
}

export const exampleMeasurementFormData: MeasurementFormData = {
    Ag: '<0.1',
    Al2O3: '15.23',
    As_: '15.5',
    Au: '0.6',
    B_ppm: '',
    Ba: '850',
    Be: '2',
    Bi: '0.2',
    CaO: '2.64',
    Cd: '0.1',
    Ce: '72',
    Co: '18.5',
    Cr2O3: '0.005',
    Cs: '6.4',
    CTOTAL___: '0.18',
    Cu: '25.6',
    Dy: '2.63',
    Er: '1.33',
    Eu: '1.18',
    Fe2O3: '7.85',
    Ga: '20.4',
    Gd: '3.46',
    geochemistry_published: true,
    Hf: '9.4',
    Hg: '0.03',
    Ho: '0.47',
    K2O: '2.63',
    La: '38.6',
    LOI: '4.8',
    Lu: '0.24',
    MgO: '1.44',
    MnO: '0.1',
    Mo: '0.7',
    Na2O: '2.95',
    Nb: '12.5',
    Nd:'29.5',
    Ni: '12.6',
    P2O5: '0.14',
    Pb: '9.8',
    Pr: '8.17',
    Rb: '89.3',
    Sb: '0.3',
    Sc: '9',
    Se: '<0.5',
    SiO2: '60.75',
    Sm: '4.9',
    Sn: '2',
    Sr: '502.2',
    STOTAL__: '0.14',
    Suma: '99.72',
    Ta: '0.9',
    Tb: '0.48',
    Te_ppm: '',
    Th: '12.7',
    TiO2: '1.23',
    Tl: '0.2',
    Tm: '0.21',
    U: '2.8',
    V: '186',
    W: '1.2',
    Y: '14.2',
    Yb: '1.64',
    Zn: '63',
    Zr: '346.2'
  }