// src/app/interfaces/measurement-payload.ts
export interface MeasurementPayload{
    CTOTAL___: number | null;
    STOTAL__: number | null;
    SiO2: number | null;
    Al2O3: number | null;
    Fe2O3: number | null;
    MgO: number | null;
    CaO: number | null;
    Na2O: number | null;
    K2O: number | null;
    TiO2: number | null;
    P2O5: number | null;
    MnO: number | null;
    Cr2O3: string;
    LOI: number | null;
    Suma: number | null;
    Sc: number | null;
    Ba: number | null;
    Be: string;
    Co: number | null;
    Cs: number | null;
    Ga: number | null;
    Hf: number | null;
    Nb: number | null;
    Rb: number | null;
    Sn: string;
    Sr: number | null;
    Ta: number | null;
    Th: number | null;
    U: number | null;
    V: number | null;
    W: number | null;
    Zr: number | null;
    Y: number | null;
    La: number | null;
    Ce: number | null;
    Pr: number | null;
    Nd: number | null;
    Sm: number | null;
    Eu: number | null;
    Gd: number | null;
    Tb: number | null;
    Dy: number | null;
    Ho: number | null;
    Er: number | null;
    Tm: number | null;
    Yb: number | null;
    Lu: number | null;
    Mo: number | null;
    Cu: number | null;
    Pb: number | null;
    Zn: number | null;
    Ni: number | null;
    As_: number | null;
    Cd: string;
    Sb: string;
    Bi: string;
    Ag: string;
    Au: string;
    Hg: string;
    Tl: string;
    Se: string;
    Te_ppm: string | null;
    B_ppm: string | null;
    geochemistry_published: boolean;
}